update test cases

+++ python/ray/_private/runtime_env.py
@@ -160,10 +160,12 @@ def create_project_package(working_dir: str, modules: List[ModuleType],
             _zip_module(module_path, module_path.parent, zip_handler)
 
 
-def delete_package_local(pkg_uri: str):
-    local_path = Path(_get_local_path(pkg_uri))
-    logger.info(f"Delete {local_path}")
-    local_path.unlink(missing_ok=True)
+def _delete_package_local(pkg_uri: str):
+    (protocol, _) = _parse_uri(pkg_uri)
+    if protocol == Protocol.GCS:
+        local_path = Path(_get_local_path(pkg_uri))
+        logger.info(f"Delete {local_path}")
+        local_path.unlink(missing_ok=True)
 
 
 def fetch_package(pkg_uri: str, pkg_file: Path) -> int:
